+++ django-rgd-imagery/tests/test_stac.py
@@ -1,0 +1,1 @@
+from rgd_imagery.stac.tests import *  # noqa
